% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{deref}
\alias{deref}
\title{Extract the value of a ref object}
\usage{
deref(x)
}
\arguments{
\item{x}{a ref object}
}
\description{
Extract the value of a ref object
}
\examples{
ref1 <- ref(1:10)
deref(ref1)
}

