% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{ref}
\alias{ref}
\title{Create a reference to R object}
\usage{
ref(object = NULL, constraint = NULL)
}
\arguments{
\item{object}{The object.}

\item{constraint}{a function to validate the
value in each update of the object.}
}
\description{
Create a ref as a wrapper of a given object.
A ref is essentially a wrapper environment of
the object.
}
\examples{
# create a NULL ref
ref()

# create a ref of an integer vector
x <- ref(1:10)
x[1]
x[5] <- 10L

# create a ref with constraint of an integer vector
x <- ref(1:10, is.integer)
x[1]
x[5] <- 10L
# x[5] <- "a" will corece 1:10 to character vector
# but violates is.integer contraint, leading to a stop.
}

