% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reflite.R
\docType{package}
\name{reflite-package}
\alias{reflite-package}
\title{The reflite package}
\description{
reflite provides a lightweight implementation of reference object.
User can 'ref' an R object (such as a vector or list) and pass it
to a function. Modifications of the object in the function body
will not make a local copy in order to mimic pass-by-refernce of
the object.
}

