% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{update.ref}
\alias{update.ref}
\title{Update the value in ref object}
\usage{
\method{update}{ref}(object, value, ...)
}
\arguments{
\item{object}{a ref object}

\item{value}{the new value}

\item{...}{additional parameters currently ignored.}
}
\description{
Update the value in ref object. The constraint (if any) will
be validated before any update of the ref object.
}
\examples{
ref1 <- ref(1:10)
update(ref1, rnorm(10))

ref2 <- ref(1:10, function(x) length(x) == 10L)
update(ref2, rnorm(10))
# The following will fail due to the violation
# of the constraint.
# update(ref2, rnorm(20))
}

